refactor(CU-#86bykk7hg): refactor code
