--- src/firstTest.test.ts
@@ -1,5 +1,0 @@
-describe('First test', () => {
-  it('should be truthy', () => {
-    expect(true).toBeTruthy();
-  });
-});
