+++ src/widgets/BlogSection/ui/AllPostSection/AllPostSection.tsx
@@ -1,13 +1,13 @@
 'use client';
 
 import React, { useMemo, useState } from 'react';
-import styles from '@/widgets/BlogSection/ui/AllPostSection/AllPostSection.module.scss';
-import container from '@/shared/styles/container.module.scss';
 import { POSTS } from '@/shared/constants/posts';
 import { PostItem } from '@/widgets/PostItem';
 import { ITEMS_PER_PAGE } from '@/widgets/BlogSection/constants/pagintation';
 import { useScrollPosition } from '@/shared/utils/hooks/useScrollPosition';
 import { useTranslations } from 'next-intl';
+import container from '@/shared/styles/container.module.scss';
+import styles from '@/widgets/BlogSection/ui/AllPostSection/AllPostSection.module.scss';
 
 export function AllPostSection() {
   const t = useTranslations('AllPostSection');
