+++ src/widgets/Header/ui/Header.tsx
@@ -2,16 +2,16 @@
 
 import React, { useState } from 'react';
 import Image from 'next/image';
-import styles from '@/widgets/Header/ui/Header.module.scss';
 import { NavigationBar } from '@/widgets/NavigationBar';
 import { NAVIGATION_LINKS_SHORT } from '@/shared/constants/navigationBar';
 import { AppTitle } from '@/shared/ui/AppTitle';
 import { BlurBackground } from '@/widgets/Header/ui/BlurBackground';
-import burgerMenuButtonSrc from '@/widgets/Header/assets/burger-menu-button-light.svg';
-import closeBurgerMenuIcon from '@/widgets/Header/assets/close-button-svgrepo-com.svg';
 import { ModalWindow } from '@/shared/ui/ModalWindow/ModalWindow';
 import { useTranslations } from 'next-intl';
 import { LocaleSwitcher } from '@/widgets/Header/ui/LocaleSwitcher/LocaleSwitcher';
+import burgerMenuButtonSrc from '@/widgets/Header/assets/burger-menu-button-light.svg';
+import closeBurgerMenuIcon from '@/widgets/Header/assets/close-button-svgrepo-com.svg';
+import styles from '@/widgets/Header/ui/Header.module.scss';
 
 export function Header() {
   const t = useTranslations('Header');
@@ -64,9 +64,7 @@ export function Header() {
         <iframe
           className={styles.videoContainer}
           title="video"
-          // src="https://www.youtube.com/embed/vDkjDmnjKC0?autoplay=1"
           src="https://www.youtube.com/embed/vDkjDmnjKC0"
-          // allow="autoplay;"
           allowFullScreen
           data-testid="video-frame"
         />
