+++ src/widgets/NavigationBar/ui/NavigationBar.tsx
@@ -1,16 +1,8 @@
 import React from 'react';
-import styles from '@/widgets/NavigationBar/ui/NavigationBar.module.scss';
-import { useTranslations } from 'next-intl';
 import Link from 'next/link';
-
-export type NavigationLinksType = {
-  text: string;
-  link: string;
-};
-
-export type NavigationBarPropsType = {
-  navigationLinks: NavigationLinksType[];
-};
+import { useTranslations } from 'next-intl';
+import { NavigationBarPropsType } from '@/shared/types';
+import styles from '@/widgets/NavigationBar/ui/NavigationBar.module.scss';
 
 export function NavigationBar({ navigationLinks }: NavigationBarPropsType) {
   const t = useTranslations('NavigationBar');
