+++ src/widgets/SpecialPostSection/ui/SpecialPostSection.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import Image from 'next/image';
-import styles from '@/widgets/SpecialPostSection/ui/SpecialPostSection.module.scss';
-import container from '@/shared/styles/container.module.scss';
 import { SpecialPost } from '@/widgets/SpecialPostSection/ui/SpecialPost';
 import specialPostImageSrc from '@/widgets/SpecialPostSection/assets/close-up-photography-of-man-wearing-sunglasses-1212984.jpg';
+import container from '@/shared/styles/container.module.scss';
+import styles from '@/widgets/SpecialPostSection/ui/SpecialPostSection.module.scss';
 
 export function SpecialPostSection() {
   return (
