+++ src/widgets/StartUpSection/ui/StartUpSection.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
+import Link from 'next/link';
 import Image from 'next/image';
-import styles from '@/widgets/StartUpSection/ui/StartUpSection.module.scss';
-import startupBackground from '@/widgets/StartUpSection/assets/startupBackground.webp';
 import { Button } from '@/shared/ui/Button/Button';
 import { POSTS } from '@/shared/constants/posts';
-import Link from 'next/link';
 import { useTranslations } from 'next-intl';
+import startupBackground from '@/widgets/StartUpSection/assets/startupBackground.webp';
+import styles from '@/widgets/StartUpSection/ui/StartUpSection.module.scss';
 
 export function StartUpSection() {
   const t = useTranslations('StartUpSection');
